+++ common/shared/src/main/java/com/chronodawn/tags/ModItemTags.java
@@ -17,6 +17,7 @@ public final class ModItemTags {
     public static final TagKey<Item> CHRONO_BOVINE_MILKING_BUCKETS = itemTag("chrono_bovine_milking_buckets");
     public static final TagKey<Item> GRASS_SEED_REPLACEMENTS = itemTag("grass_seed_replacements");
     public static final TagKey<Item> PORTAL_IGNITERS = itemTag("portal_igniters");
+    public static final TagKey<Item> PORTAL_STABILIZERS = itemTag("portal_stabilizers");
 
     private ModItemTags() {
     }